% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFS85.R
\name{MFS85}
\alias{MFS85}
\title{ALL Metabolic Subtypes}
\arguments{
\item{expr.data}{data.frame. Genes expression data.}

\item{gsva.methods}{character. The method of GSVA to estimate gene-set enrichment scores per sample. See \link{gsva}}

\item{ncores}{numeric. Number of threads used for GSVA analysis. See \link{gsva}}

\item{scale}{logical. Whether to scale the genes expression data.}
}
\value{
result
}
\description{
This R software package used GSVA to calculate the enrichment abundance of 85 metabolic pathways through transcriptome data of new samples, and based on the metabolic-feature-based subtypes (MFSs) and enrichment abundance of 85 metabolic pathways of the built-in training set, the new samples were classified as MFS1, or MFS2 or MFS3.
}
